Move tower_common, fragment, vs modules under lib/ and update imports
