+++ lib/domains/tower/tower_common.py
